Cast every double to the type of qreal in qBound calls.

+++ 3rdparty/posterazor/posterazorcore.cpp
@@ -265,22 +265,22 @@ QPrinter::Orientation PosteRazorCore::paperOrientation() const
 
 double PosteRazorCore::paperBorderTop() const
 {
-    return qBound(.0, convertCmToDistance(m_paperBorderTop), maximalHorizontalPaperBorder());
+    return qBound((qreal).0, convertCmToDistance(m_paperBorderTop), maximalHorizontalPaperBorder());
 }
 
 double PosteRazorCore::paperBorderRight() const
 {
-    return qBound(.0, convertCmToDistance(m_paperBorderRight), maximalVerticalPaperBorder());
+    return qBound((qreal).0, convertCmToDistance(m_paperBorderRight), maximalVerticalPaperBorder());
 }
 
 double PosteRazorCore::paperBorderBottom() const
 {
-    return qBound(.0, convertCmToDistance(m_paperBorderBottom), maximalHorizontalPaperBorder());
+    return qBound((qreal).0, convertCmToDistance(m_paperBorderBottom), maximalHorizontalPaperBorder());
 }
 
 double PosteRazorCore::paperBorderLeft() const
 {
-    return qBound(.0, convertCmToDistance(m_paperBorderLeft), maximalVerticalPaperBorder());
+    return qBound((qreal).0, convertCmToDistance(m_paperBorderLeft), maximalVerticalPaperBorder());
 }
 
 double PosteRazorCore::maximalVerticalPaperBorder() const
@@ -429,12 +429,12 @@ void PosteRazorCore::setOverlappingHeight(double height)
 
 double PosteRazorCore::overlappingWidth() const
 {
-    return qBound(.0, convertCmToDistance(m_overlappingWidth), maximalOverLappingWidth());
+    return qBound((qreal).0, convertCmToDistance(m_overlappingWidth), maximalOverLappingWidth());
 }
 
 double PosteRazorCore::overlappingHeight() const
 {
-    return qBound(.0, convertCmToDistance(m_overlappingHeight), maximalOverLappingHeight());
+    return qBound((qreal).0, convertCmToDistance(m_overlappingHeight), maximalOverLappingHeight());
 }
 
 double PosteRazorCore::maximalOverLappingWidth() const
@@ -736,13 +736,13 @@ void PosteRazorCore::paintPosterPageOnCanvas(PaintCanvasInterface *paintCanvas, 
         :alignment & Qt::AlignHCenter?(posterTotalSizeCm.width() - posterImageSizeCm.width())/2 - borderLeftCm
         :-borderLeftCm
     );
-    imageOffsetFromLeftPosterBorderCm = qBound(.0, imageOffsetFromLeftPosterBorderCm, posterTotalSizeCm.width() - posterImageSizeCm.width() - borderLeftCm - borderRightCm);
+    imageOffsetFromLeftPosterBorderCm = qBound((qreal).0, imageOffsetFromLeftPosterBorderCm, posterTotalSizeCm.width() - posterImageSizeCm.width() - borderLeftCm - borderRightCm);
     double imageOffsetFromTopPosterBorderCm = (
         alignment & Qt::AlignBottom?posterTotalSizeCm.height() - posterImageSizeCm.height() - borderTopCm
         :alignment & Qt::AlignVCenter?(posterTotalSizeCm.height() - posterImageSizeCm.height())/2 - borderTopCm
         :-borderTopCm
     );
-    imageOffsetFromTopPosterBorderCm = qBound(.0, imageOffsetFromTopPosterBorderCm, posterTotalSizeCm.height() - posterImageSizeCm.height() - borderTopCm - borderBottomCm);
+    imageOffsetFromTopPosterBorderCm = qBound((qreal).0, imageOffsetFromTopPosterBorderCm, posterTotalSizeCm.height() - posterImageSizeCm.height() - borderTopCm - borderBottomCm);
     const QPointF pageOffsetToImageFromTopLeftCm(
         column * (printablePaperAreaSizeCm.width()- overlappingWidthCm) - imageOffsetFromLeftPosterBorderCm,
         row * (printablePaperAreaSizeCm.height() - overlappingHeightCm) - imageOffsetFromTopPosterBorderCm
